Misc doc (#156)
- Make helper macros' visibility crate-private
- Update some broken doc links
- Fix typos
