+++ assembly/src/execution/trace.rs
@@ -96,9 +96,9 @@ pub struct BoundaryValues {
 /// Convenience macro to execute all the flushing rules of a given kind of
 /// instructions present in a [`PetraTrace`].
 ///
-/// It takes as argument the list events for the targeted instruction in a
-/// trace, the [`InterpreterChannels`] against which the flushing rules will be
-/// performed, and the [`InterpreterTables`].
+/// It takes as argument the list of events for the targeted instruction in a
+/// trace and the [`InterpreterChannels`] against which the flushing rules will
+/// be performed.
 ///
 /// # Example
 ///
